simplify scss management
